requirements.txt
asgiref==3.4.1
certifi==2021.5.30
chardet==4.0.0
Django==3.2.5
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
idna==2.10
importlib-metadata==4.6.1
Markdown==3.3.4
psycopg2-binary==2.9.1
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.6
zipp==3.5.0
